Bubble_sort
Currently working with a smaller array to test conditionals, loops, etc.
as a starting point. The #sweep method uses a while loop to check if an
element is greater than the one before it. If it is, it will swap those
two elements. After that method has finished its loops and if..else
block checks if [0] < [1] and [1] < [2]. If this is not true is will run
the #sweep method again.

Even with working with three elements in the test array it isn't working
correctly. Negative numbers also confuse the method, though it's not
part of the assignment.