Sweep and check methods
This commit changes the sweep method and adds a check method.

The sweep method now compares [i-1] > [i] instead of [i] > [i+1]. This
may have caused the return array to have an extra nil when trying to
sort the test array.

The check method is meant to check the elements in the array after the
intial sweep. If an element, i, is less than i-1, it will call the sweep
method again, else it will just break.

Currently the return array isn't fully sorted in order; the only issue
is in the first two elements. [2, 0, 2, 3, 4, 78]