Had to check solutions to complete...
This assignment was difficult and I resorted to looking at another's
code to understand how to solve it. Rather than just simply copying the
code, I tried to understand line-by-line what was happening and tried to
reproduce it again. There were still some parts I got wrong and was
unable to reproduce the same results.

The solution that I saw used a while loop that was embedded in another
while loop. The innermost while loop would keep looping until 'i', the
last element in the array is put through the comparison if statement
(i is increased by 1 after each if block). After each if statement is
run, the 'unsorted' variable is assigned 'true'.

Once the code is out of the inner while loop, the outer while loop is
conditioned to keep running while 'unsorted' is true. Since 'unsorted'
has been assigned true, the next two lines that are run change the value
of i to 0 and 'unsorted' to false (which means the outer while loop's
condition is no longer met). By now the array is sorted and printed.